Updated GUI
Added restart button, more menu options(optimal ordering and clear row currently drop pentominoes according to our needs but we can make an animation of it). Improved look(maybe when Stijn finishes GameBoard we can make a screenshot of it and use it as a background for the other windows). Highscore works and updates, if no file is available to read from highscore will return no data available but after the first game a new file will be created. Additional things are there too... Also three of the pentominoes are with mirrored coordinates so we can use them in optimal ordering Basia has.